wip: add setup files. ready for testing
