+++ 5.35.0/index.ts
@@ -4,13 +4,14 @@ import { updateToEmotion11 } from "./updateToEmotion11";
 import { updateAdminApp } from "./updateAdminApp";
 import { setupFiles } from "./setupFiles";
 import { Context } from "../types";
+import {migrateThemeTypography} from "./migrateThemeTypography";
 
 module.exports = async (context: Context) => {
     const files = setupFiles(context);
     const rawFiles = files.paths();
     const project = createMorphProject(rawFiles);
 
-    await [updateGraphQL, updateToEmotion11, updateAdminApp].reduce(async (_, processor) => {
+    await [updateGraphQL, updateToEmotion11, updateAdminApp, migrateThemeTypography].reduce(async (_, processor) => {
         return await processor({
             context,
             project,
