+++ 5.35.0/themeTypographyMigration/migrateImports.ts
@@ -2,7 +2,7 @@ import { SourceFile } from "ts-morph";
 import {
     ImportDeclarationDefinition,
     ThemeFileMigrationDefinition
-} from "./migrationFileDefinitions";
+} from "./themeMigrationSetupFiles";
 import { StyleIdToTypographyTypeMap } from "./definitions";
 import { Context } from "../../types";
